Port emacs-config.el to emacs-config.org; also add init.el
We have ported the emacs-config.el file to emacs-config.org,
which is an org-mode file with emacs-lisp code snippets included.

An init.el is also provided.  Both init.el and emacs-config.org are
copied to the ~/.emacs.d folder when running the install.sh script.

init.el org-babel-load-file to load the emacs-config.org on startup.
Configuration settings are in emacs-lisp code blocks that will get
executed.  An emacs-config.el file corresponding to the settings
contained in emacs-config.org will be created each time you
start emacs.

Also removed the sample-init-el folder; that is no longer needed.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>